requirements.txt
aiosmtpd==1.4.6
alembic==1.15.1
atpublic==5.1
attrs==25.3.0
blinker==1.9.0
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
cssbeautifier==1.15.4
djlint==1.36.4
dnspython==2.7.0
EditorConfig==0.17.0
elastic-transport==8.17.1
elasticsearch==8.17.2
email_validator==2.2.0
firebase-admin==6.7.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Mailman==1.1.1
Flask-Migrate==4.1.0
Flask-Moment==1.0.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsbeautifier==1.15.4
json5==0.10.0
Mako==1.3.9
Markdown==3.7
MarkupSafe==3.0.2
msgpack==1.1.0
packaging==24.2
pathspec==0.12.1
proto-plus==1.26.1
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyotp==2.9.0
pyparsing==3.2.1
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
qrcode==8.0
redis==5.2.1
regex==2024.11.6
requests==2.32.3
rq==2.3.1
rsa==4.9
six==1.17.0
SQLAlchemy==2.0.39
text-unidecode==1.3
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
WTForms==3.2.1
